rscore: close the coverage gaps the parity gate could not see
Audit findings, each verified against the code before fixing.

Coverage, not hashing, was the remaining hole:
- assertShadowParity returned silently while the mirror was still booting
  (dynamic import, Hello, Restore), so a short strict run could pass having
  compared nothing. It now waits for arming and for the queue to drain, and
  an unarmed mirror fails strict outright.
- The gap hook is armed before the buffered frames replay; a gap in that
  first batch used to miss the halt entirely.
- An account seeded from committed TypeScript state reproduces the TS root
  by construction. Seeded accounts that never executed a transition are
  tracked and fail the gate, and executions are counted by tx type.
- Owners beyond the binding limit never reach the engine; skippedUnboundOwner
  now fails the gate instead of being invisible next to a green owner.

Correctness:
- The receiver path executed with the frame's own clock instead of the
  Entity enforcement clock TypeScript used, so a legitimately accepted late
  timeout looked like divergence. It now receives securityContext's clock.
- The output comparison covered a forward's fields and a secret's lock id.
  It now covers delivery mode, the full released-secret envelope (hashlock,
  secret, token, amount) and the timeout outcome, projected from the
  authority's own per-tx results rather than re-derived from the txs.
- policyVersion is u64 in the engine: TypeScript admits any positive safe
  integer, and the u32 decoder disabled the mirror instead of comparing.
- The wire fingerprint moves with the wire shape (wire=2), so an engine
  built for the old seed/summary layout is rejected at Hello.
- Requests carry a deadline; a wedged engine fails the run instead of
  hanging every caller waiting on the mirror.

+++ rscore/crates/engine/src/error.rs
@@ -78,7 +78,7 @@ pub enum ValidationRejection {
         token_id: u32,
     },
     RebalancePolicyVersion {
-        version: u32,
+        version: u64,
     },
     RebalancePolicyTimestamp,
     RebalancePolicyFeeTerms {
@@ -90,7 +90,7 @@ pub enum ValidationRejection {
     RebalancePolicyEquivocation {
         side: Side,
         token_id: u32,
-        version: u32,
+        version: u64,
     },
     Htlc(HtlcRejection),
 }
